separate map viewport and bounds hooks
